[jwies/review-budget-cap] review: split the budget table into budgets.ts (router.ts over the max-lines limit after the main merge)
