+++ workflows/review/lib/router.ts
@@ -30,6 +30,7 @@
  * sentence about the change itself. Prose stays with the lens sub-agents.
  */
 
+import {DEFAULT_MISROUTED_FLOOR_TIER, DEFAULT_TIER_BUDGETS} from "./budgets";
 import {clampBudgetToCreditCap, resolveCreditCap} from "./credit-cap";
 import {KNOWN_LENSES} from "./finding-schema";
 import type {Lens} from "./finding-schema";
@@ -49,6 +50,7 @@ import type {
 
 // Re-exported so consumers (and the tests) can treat the router as the single
 // entry point for routing vocabulary and the ROUTING parser.
+export {DEFAULT_MISROUTED_FLOOR_TIER, DEFAULT_TIER_BUDGETS};
 export {
     ENABLEABLE_REVIEWERS,
     parseRoutingConfig,
@@ -277,81 +279,6 @@ export type RouteInput = {
      */
     resolvedTiers?: Record<string, RiskTier>;
 };
-
-/* -------------------------------------------------------------------------- */
-/* Documented defaults (tunable later)                                        */
-/* -------------------------------------------------------------------------- */
-
-/**
- * Default budget table, sized inside the workflow's assumed 20-minute / $10
- * per-run ceiling. Every field scales monotonically with the tier, and the
- * table is exported so the eval suite and consumers can override it without a
- * code change.
- *
- * Calibration (re-measured 2026-07-10 against production runs on
- * Khan/actions #232/#238): a run spends ~3 minutes of fixed overhead
- * (staging, router, provenance, pattern-triage) before the first reviewer
- * returns, so wall clocks below ~6 minutes put a run past its shed threshold
- * before any review work lands; and the standard enabled roster is seven
- * whole-change reviewers (two defaults plus the five eval-justified opt-ins
- * both current consumers enable), so an invocation cap of 4 deterministically
- * shed dimensions on every low-tier run. Low and medium now fit the roster;
- * trivial stays deliberately small (a trivial-only PR gets the two default
- * reviewers, with the rest declared as budget sheds). Raising the low cap to
- * 8 is deliberate: a lens-free low-tier run can dispatch the full roster.
- * Path-matched specialist lenses also consume slots, so a low PR matching
- * two or more lenses still sheds from the bottom of the value ranking; that
- * residual is sized when the modes are priced, not here. `maxUsd` is
- * deliberately left uncalibrated
- * (per-run cost measurement is deferred to #249), so the dollar column is
- * still the original estimate, not a re-measured figure.
- */
-export const DEFAULT_TIER_BUDGETS: Record<RiskTier, RunBudget> = {
-    trivial: {
-        tier: "trivial",
-        floored: false,
-        maxReviewerInvocations: 2,
-        maxToolCallsPerFinding: 2,
-        maxTotalToolCalls: 10,
-        maxWallClockMinutes: 6,
-        maxUsd: 0.5,
-    },
-    low: {
-        tier: "low",
-        floored: false,
-        maxReviewerInvocations: 8,
-        maxToolCallsPerFinding: 3,
-        maxTotalToolCalls: 20,
-        maxWallClockMinutes: 10,
-        maxUsd: 1.5,
-    },
-    medium: {
-        tier: "medium",
-        floored: false,
-        maxReviewerInvocations: 10,
-        maxToolCallsPerFinding: 5,
-        maxTotalToolCalls: 60,
-        maxWallClockMinutes: 15,
-        maxUsd: 4,
-    },
-    high: {
-        tier: "high",
-        floored: false,
-        maxReviewerInvocations: 12,
-        maxToolCallsPerFinding: 8,
-        maxTotalToolCalls: 120,
-        maxWallClockMinutes: 20,
-        maxUsd: 10,
-    },
-};
-
-/**
- * Tier a misrouted PR (source files touched, but no specialist lens matched) is
- * floored to. A misrouted PR still gets a real review from the always-on
- * reviewers, so it must not fall to the trivial budget just because no path
- * pattern claimed it. "low" is the documented default floor.
- */
-export const DEFAULT_MISROUTED_FLOOR_TIER: RiskTier = "low";
 
 /* -------------------------------------------------------------------------- */
 /* Glob matching (self-contained; CODEOWNERS/gitattributes-style)            */
